Refactor managePassThrough script: extract transferTokens function and update address mappings
