+++ contracts/scripts/defender-actions/managePassThrough.js
@@ -1,12 +1,8 @@
-const { ethers } = require("ethers");
 const {
   DefenderRelaySigner,
   DefenderRelayProvider,
 } = require("@openzeppelin/defender-relay-client/lib/ethers");
-const addresses = require("../../utils/addresses");
-
-const passThroughAbi = require("../../abi/passThrough.json");
-const { logTxDetails } = require("../../utils/txLogger");
+const { transferTokens } = require("../../utils/managePassThrough");
 
 // Entrypoint for the Defender Action
 const handler = async (event) => {
@@ -18,43 +14,7 @@ const handler = async (event) => {
   const provider = new DefenderRelayProvider(event);
   const signer = new DefenderRelaySigner(event, provider, { speed: "fastest" });
 
-  const OUSD = addresses.mainnet.OUSDProxy;
-  const OETH = addresses.mainnet.OETHProxy;
-
-  // Map tokens to their passThrough contracts
-  const tokenPassThroughs = {
-    [OUSD]: [
-      "0x261Fe804ff1F7909c27106dE7030d5A33E72E1bD", // OUSD/3pool Curve pool
-      "0xF29c14dD91e3755ddc1BADc92db549007293F67b", // OUSD/USDT  Uniswap pool
-    ],
-    [OETH]: [
-      "0x2D3007d07aF522988A0Bf3C57Ee1074fA1B27CF1", // OGN/OETH   Uniswap pool
-      "0x216dEBBF25e5e67e6f5B2AD59c856Fc364478A6A", // OETH/WETH  Uniswap pool
-    ],
-  };
-
-  console.log("DEBUG: Token PassThroughs mapping", tokenPassThroughs);
-
-  // Process all tokens and their passThrough contracts
-  for (const [token, passThroughAddresses] of Object.entries(
-    tokenPassThroughs
-  )) {
-    const tokenName = token === OUSD ? "OUSD" : "OETH";
-
-    for (const passThroughAddress of passThroughAddresses) {
-      const passThrough = new ethers.Contract(
-        passThroughAddress,
-        passThroughAbi,
-        signer
-      );
-
-      const tx = await passThrough.connect(signer).passThroughTokens([token]);
-      await logTxDetails(
-        tx,
-        `PassThrough at ${passThroughAddress} sent ${tokenName}`
-      );
-    }
-  }
+  await transferTokens({ signer });
 };
 
 module.exports = { handler };
